move prject files
