--- tailwind.config.ts
@@ -1,34 +1,0 @@
-import type { Config } from "tailwindcss";
-
-const config: Config = {
-  darkMode: "class",
-  content: [
-    "./pages/**/*.{ts,tsx}",
-    "./components/**/*.{ts,tsx}",
-    "./app/**/*.{ts,tsx}",
-    "./src/**/*.{ts,tsx}",
-    "*.{js,ts,jsx,tsx,mdx}",
-  ],
-  prefix: "",
-  theme: {
-    container: {
-      center: true,
-      padding: "2rem",
-      screens: {
-        "2xl": "1400px",
-      },
-    },
-    extend: {
-      colors: {
-        background: "#F1F8E8",
-        lightBlue: "#80ACD1",
-        foobar: "#202E0E",
-        darkGreen: "#283A12",
-        gold: "#7B783D",
-        green: "#304616",
-      },
-    },
-  },
-};
-
-export default config;
