+++ travel-blog/app/posts/[slug]/page.tsx
