+++ travel-blog/components/Badge.tsx
