+++ travel-blog/components/BlogPostCard.tsx
