+++ travel-blog/components/Card.tsx
