+++ travel-blog/components/CardContent.tsx
