+++ travel-blog/components/HeroSection.tsx
